requirements.txt
requests>=2.25.0
pandas>=1.3.0
matplotlib>=3.5.0
plotly>=5.0.0
streamlit>=1.25.0
python-docx
reportlab
